Make bindingPackage a concrete method on Binding instead of requiring all binding value types to define it.
Every case of findBindingPackage() was always passing it the value which would eventually be stored in key(), except for bindings resulting from Components, Component methods, and Subcomponent builders. For each of those cases, we know the type will be accessible from the component since it is already referenced in the component declaration file and raw types won't be needed, just as before.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=124366368