Compact classnames for subcomponent implementations, fixes #421.
Fix for https://github.com/google/dagger/issues/421.

Instead of including more and more levels of container names until we achieve
uniqueness, we try two strategies: use simple names everywhere, or else
include a uniquing prefix everywhere.

The prefix is built from the initials of a containing class for nested
subcomponent declarations, or the initials of containing packages for
top level. An integer may appended to the prefix if necessary.