Automated rollback of 09ed0bba
*** Original change description ***

Fix a bug in the way that `@Binds` methods are attributed to their parent components when they are unused in their parent.

I believe the only practical way this would have come up is for `@ProductionScope` bindings in `@ProductionComponent`s and `@ProductionSubcomponent`s since it can be applied to many components, but it's possible there are other cases. I wasn't able to reproduce this with a regular scope+regular components.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=169167410