Memoize the hash code of BindingGraph
This solves a major performance regression for creating a BindingNetwork, which used a List<BindingGraph> in one of its node types. Creation of a BindingNetwork dropped from 84 seconds to under 2 seconds during local profiling on my machine for the largest component I know of. This should allow us to make progress in implementing the Dagger SPI.

This should also have performance improvements broadly, initial timings seem to be in the 10-30% range depending on the shape of a component.

We've long suspsected that BindingGraph's hashCode was a bottleneck but this seems like the best evidence we've seen. I imagine this won't do much for compilations that only have a single component, but in the situations where subcomponents are used (and we have an ImmutableSet<BindingGraph>) there should be good gains.

RELNOTES=Improved compilation performance of BindingGraph

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=175729695