Pass variant javacopts to Dagger's functional tests.
We were only passing the variant javacopts to the libraries of those functional tests. However, that silently fails to apply the javacopts when the @Component is defined in the test class itself, since the javacopts were not applied to the test classes.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=193042437