Use static factory methods for Producer factories to bring them in line with how we create regular provision factories
In addition to the benefits of consistency, there is a practical reason for this: when we run in hjar mode, we only have a special codegen strategy for components, but not factories. I'd like to add a new strategy for all source generators that does filtering of the JavaPoet models - this will bring hjar mode to all of our codegen without any special configuration for each generator. This currently doesn't work because Turbine needs (for some technical reasons in javac) super() calls for constructors to remain present, and filtering a MethodSpec.code for just a super() call is non-trivial. This change makes the constructors private, and thus are filterable for Turbine entirely (sidestepping+ignoring the super() call).

This strategy will be critical when we start implementing hjar support for ahead-of-time subcomponents, since the generated subcomponents expose protected APIs to further compilations, and thus the code we must generate in Turbine is much more involved than it currently is.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=217314566