Simplify the binding types of dagger.android injector factories
We've removed the need for "extends FrameworkType" in the return type and are converging on @ClassKey instead of all of the different map keys for each framework type. This will allow us to easily add support for new AndroidX types in dagger.android in the future.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=218341740