Move producer cancellation out of ProducerNodeInstanceBindingExpression and into ComponentModelBuilder
Without this change, producer cancellation statements always refer to their direct field, but in ahead-of-time subcomponents, these may be wrapped in a method and later modified. Moving this into ComponentModelBuilder allows to break an infinite recursion and reference a binding expression for the producer directly, even if it's wrapped in a method.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=219185296