When modifiable framework instances are in fact modified, ensure that superclass implementations get the correct provider instance during initialization.
DelegateFactory is used to solve this cycle (though not a binding cycle, it is a "codegen" cycle of sorts), and a DelegateProducer is added for the same purpose but for producers.

In order for the delegate factories to be initialized properly, initialization is moved out of the constructors of abstract subcomponent implementations and into a method that can be called after super(); calls in the final implementation.

This is a rollforward of 218dd448dfd1d791480f13e507eed567fdabae43, which relied on Guava's functional types extending the java.util.function types (which continue to bite us in bizarre ways). I've verified that the broken targets now build

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=221828464