Simplify ComponentBindingExpressions.wrapInMethod() by reducing the parameters needed for MethodBindingExpression, inlining some of the logic about when to use a modifiable method, and switching to an imperative style instead of a functional style.
I think it's much easier to follow when and why we use a modifiable method now vs. before.

Also rename matchingModifiableBindingMethod to supertypeModifiableBindingMethod. I think this is a clearer explanation of what this is - it just so happens that at the time that we call getModifiableBindingMethod() in a subclass implementation that we haven't added it to this ComponentImplementation, but the previous name made it seem like it may be a a matching modifiable method in _this_ component, which it isn't.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=222444932