Run binding graph validation on subcomponents that are being processed in ahead-of-time subcomponents mode
In doing so, I uncovered and fixed a bug in binding graph resolution for scoped @Inject constructors in AOT. When we normally resolve an @Inject binding, if there is no component that shares the scope, we resolve the binding at the root. But that shouldn't happen in a partial graph, where the @Inject binding could be owned by a parent that *does* have the scope.

This isn't a problem for unscoped bindings because they could be owned effectively anywhere.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=225586785