Report duplicate bindings even if they have no incoming dependency requests. This is in preparation for module binding validation, where there may not be any incoming dependency requests for a binding.
Reports only a warning if the duplicate is an explicit binding that conflicts with an @Inject binding in an ancestor component. This was never intended to be supported but never checked. Adds a flag to control whether that is a warning or an error.

RELNOTES=Reports a warning if an explicit binding conflicts with an `@Inject` binding in an ancestor component. This was never intended to be supported but never checked. Adds a flag to control whether that is a warning or an error.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=227751023