Propagate compiler options related to generation from a base implementation through to all of its future subclass implementations
This is necessary for options like fastInit, which may generate a different set of modifiable methods if it is enabled vs. disabled. If we only use the flags that are passed to the current compilation, our superclass model will be thrown off. It's also probably important for consistency of some other flags

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=229269597