Fix the @GenerationOptions retrieval and enable a build_test for the dagger.functional.aot.fastinit libraries so that we catch this in presubmit.
I originally had this and dpb@ asked what it was doing, and I wasn't able to remember why I had done it that way, and all of the tests were passing so I "simplified" it. I added a comment about what's going on so it's clearer in the future.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=229463931