For non-abstract modules that A) are required by a component, B) can be instantiated by Dagger, and C) have no corresponding setter in the component's builder, do not add a field to the builder for the module.
Instead, instantiate the module inline when calling the component's constructor.

The main impetus for this (besides getting rid of the weirdness of having a field for something the user can never set) is to ensure that the behavior of builders with no setter methods and factories (once factories are usable) is the same. A factory will definitely never have any state.

This is a potential behavioral change because before, such a module would be cached in the builder after first being created and if multiple component instances were created from the same builder, they would all use the same instance of the module. Now, each component will get a different instance of the module.

That said, the situation in which an observable behavioral change could occur should be vanishingly rare. The following conditions would all need to be true:

- A component depends on a Dagger-instantiable concrete module that has mutable instance state.
- The component has a builder that does not provide a setter for that module.
- A single instance of that builder is used to construct multiple component instances.

This behavior was never guaranteed, most people should not be using a builder to construct multiple component instances, and anyone relying on module instance state (something that should be avoided in general) should really be providing a setter to explicitly set the module instance to use.

RELNOTES=When a single instance of a component builder with no setter for a particular module that Dagger can instantiate is used to instantiate multiple component instances, provide a new instance of that module to each component rather than caching and reusing a single instance of the module for each component.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=229576758