Rename SUBCOMPONENT_BUILDER enum constants to SUBCOMPONENT_CREATOR and rename some SubcomponentBuilder types to SubcomponentCreator.
The dagger.model names being changed here don't seem to be used outside of dagger packages; it's possible that they're used externally, but I'm not sure how to change the BindingKind in a way that allows for migration.

RELNOTES=Renamed `BindingKind.SUBCOMPONENT_BUILDER` to `SUBCOMPONENT_CREATOR` to reflect the fact that it can be a builder or a factory, and renamed the edge type to `SubcomponentCreatorBindingEdge`.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=231224052