Don't recompute top-level ComponentImplementation instances if we already have created one in the same round.
This reduces the cost of compiling multiple subcomponents that participate in the same ancestry together. It's still likely to be better to split those out into separate compilations, but there shouldn't be an additional tax if you don't.

Also define a `ClearableCache` interface as this is now the third instance of a caching factory that we want to purge across processing rounds.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=233959368