Refactor ModuleBindingValidationTest some more to prepare for applying full binding graph validation to components and subcomponents.
Move tests/constants around and rename so it's clearer which files/messages go with which tests.

Add back the case where a module includes another module with errors/warnings to demonstrate that we don't repeat the error (but we do repeat the warning: see b/130284666).

Add a similar case where a module declares a subcomponent with errors/warnings to demonstrate that we DO repeat the error/warning (see b/130284677).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=242910728