Extract a BindingElementValidator type from BindingMethodValidator
Add a BindsInstanceValidator to validate some of the common things that it shares with other bindings. This fills a gap of validation, but also will make adding multibinding annotations to @BindsInstance declarations simpler.

An alternative to this CL is to expose some of the methods from BMV as static methods elsewhere, and then allow a BindsInstance validator to call them. I think I prefer the BindingElementValidator approach better as it makes it harder to have skew between the two in the future.

RELNOTES=`@BindsInstance` annotations are now more strictly validated

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=246172335