Add blaze rule to validate and generate files for a maven artifact.
This CL adds a skylark macro, `gen_maven_artifact`, that generates all of the maven artifact files:

  * pom.xml
  * jar file
  * src jar file
  * javadoc jar file

In addition, the macro validates that the user has specified all of the target's required dependencies. A dependency is "required" if it is not included in any of the maven dependencies of the target.

RELNOTES=N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=272729031