Add version tracking file to deployed Dagger artifacts.
This CL adds logic to the deploy scripts so that a file is added to the
META-INF folder of the jars. The file will then contain the version name
of the library so that we can track APKs that use Dagger. The name of
the file is '<group-id>_<artifact-id>.version'.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=286078333