Remove usages of Guava APIs not available in the '-android' variant.
Even though Gradle has separate processor and app classpath it is
common for users to end up using the `-android` variant of Guava in
the processor part which causes Dagger to fail when using Guava APIs for
Java 8. To avoid we are adding the `-android` variant dependency in a
Gradle sample as a smoke test.

Fixes #1700

RELNOTES=Remove usages of Guava APIs not available in the '-android' variant.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=288497626