Discover companion object bindings of @Module annotated classes.
If a Kotlin class annotated with @Module has a companion object, then
the functions in it are considered as part of the bindings declared by
the module. The bindings are effectively static and don't require an
instance. For backwards compatibility, it is OK to still annotated the
companion object with @Module and its functions with @JvmStatic, but
this is no longer needed.

RELNOTES=Support binding declarations in Kotlin companion objects of @Module annotated classes.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=288711739