Move identifiernamestring to proguard config + enable default
This moves the `identifiernamestring` directly to the shipped proguard config. R8 is the default now, so this should be safe to enable by default at this point and I've updated the flag to match.

This is important also because the current implementation was not compatible with incremental annotation processing, and would force full rebuilds if stringKeys() was enabled. This was because it did not report originating elements. Since it's also a hardcoded construct and not attached to a particular element, it would have nothing to originate from even if we did want to add them.