Move dagger-android to androidx
Resolves #1652
Resolves #1619
Resolves #1248
Resolves #1489
Resolves #1444
Resolves #1313
Resolves #1264

Enables better support for newer androidx subprojects like #1271, but I feel we need to start somewhere. This unblocks that by making the initial move to androidx as a namespace, and then splitting up newer artifacts out of this (and progressively making dagger-android-support increasingly just a shim that delegates to them) is a healthy approach.