Move identifiernamestring to proguard config.
This moves the `identifiernamestring` directly to the shipped proguard
config. R8 is the default now, so this should be safe to enable by
default at this point but the flag is not yet updated in this change.

This change is important also because the current implementation was not
compatible with incremental annotation processing, and would force full
rebuilds if stringKeys() was enabled. This was because it did not report
originating elements. Since it's also a hardcoded construct and not
attached to a particular element, it would have nothing to originate
from even if we did wanted to add them.

Closes https://github.com/google/dagger/pull/1720

RELNOTES=Generate 'identifiernamestring' r8 rule by default.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=291056440