Upstream DaggerKotlinIssuesDetector
This is a set of custom lints we wrote internally to help with migration to improved kotlin support in Dagger.

This covers the following cases:
- Using `field:` site target in injected properties
- Using `@JvmStatic` in object modules
- Annotating `companion object`s with `@Module`

These checks also provide lintFixes.

One significant caveat - the lint jar must be applied as an artifact directly, as the bazel tooling for `android_library` is quite lacking and doesn't offer an equivalent to Android Gradle Plugin's `lintPublish`.

For Android consumers, using this would look like so:

```gradle
dependencies {
  lintChecks "com.google.dagger:dagger-lint:<version>"
}
```

For plain Kotlin projects, they can use the lint plugin (I believe this is right, but it's been a minute since I've dealt with it)

```gradle
apply plugin: "com.android.lint"

dependencies {
  lintChecks "com.google.dagger:dagger-lint:<version>"
}
```