Skip dependency verification if project is already failing.
When the Gradle project has a failure it might not have its dependencies
resolved, therefore we should skip verifying the dependencies or else we
might spam the error log with misleading error.

This can happen for example if a dependency is wrongly typed or a plugin
is misconfigured.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=306331358