Roll forward (try 2): Organize Hilt gradle example tests and update to use ActivityScenario.
New: The issue was a deadlock with the synchronization. To fix, I've removed the synchronization completely, since it was an unrelated change to the rest of the CL. I filed a bug and it will be address separately from this CL (see b/154964957).

*** Original change description ***

Roll forward: Organize Hilt gradle example tests and update to use ActivityScenario.

New: Allow TestApplicationComponentMa...

***

RELNOTES=N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=308676812