Transform super call invocations to reference generated superclass.
Iterates over each declared method, finding in its bodies super calls.
(e.g. super.onCreate()) and rewrites the method reference of the
invokespecial instruction to one that uses the new superclass.

The invokespecial instruction is emitted for code that between other
things also invokes a method of a superclass of the current class.
The opcode invokespecial takes two operands, each of 8 bit, that
together represent an address in the constant pool to a method
reference. The method reference is computed at compile-time by looking
the direct superclass declaration, but at runtime the code behaves
like invokevirtual, where as the actual method invoked is looked up
based on the class hierarchy.

However, it has been observed that on APIs 19 to 22 the Android
Runtime (ART) jumps over the direct superclass and into the method
reference class, causing unexpected behaviours. Therefore, this method
performs the additional transformation to rewrite direct super call
invocations to use a method reference whose class in the pool is the
new superclass. Note that this is not necessary for constructor calls
since the Javassist library takes care of those.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=310243862