Upstream DoNotCallProvidersDetector
This adds a lint that detects user-code calls to Dagger provider methods, which are intended to just be called by generated dagger code. We currently run this as an error, but open to making it just a warning as a (likely) code smell. I'd recommend error and allow suppression as-needed.