Add FIELD annotation target to @ActivityContext qualifier.
If the qualifiers does not have the target field then it can't
be used with in members that are injected either by entry points
or @AndroidEntryPoint.

This change also removes the @Target(TYPE) the annotation contained,
this was a mistake and there is no reason for this qualifier to be
added to a class definition.

Fixes https://github.com/google/dagger/issues/1980

RELNOTES=Added missing FIELD target to @ActivityContext qualifier.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=320698031