Prevent KAPT from logging errors when compiling modules without entry points
When Hilt Android is run on a module that contains no entry points, KAPT
will log an error about disableAndroidSuperclassValidation not being a
support option. This is because the Hilt Gradle plugin will add this
option when applied but if no entry points exists, no processor actually
accepts that option.

With this fix, all hilt processors accept these options (though they are
unused). This will prevent KAPT from logging these warnings in most cases.

It is still possible for warnings to be logged, for example if the Hilt
plugin is applied, but Hilt isn't used but for those cases the user can
simply skip applying the plugin.

This fixes #2040